--- NotebookSystem.layout
@@ -1,23 +1,0 @@
-[Editors]
-Order=0,20
-Focused=0
-[Editor_0]
-CursorCol=37
-CursorRow=328
-TopLine=297
-LeftChar=1
-[Editor_18]
-CursorCol=87
-CursorRow=290
-TopLine=256
-LeftChar=1
-[Editor_19]
-CursorCol=39
-CursorRow=211
-TopLine=151
-LeftChar=1
-[Editor_20]
-CursorCol=40
-CursorRow=49
-TopLine=9
-LeftChar=1
