+++ global.h
@@ -1,5 +1,5 @@
-#ifndef _MAIN_H_
-#define _MAIN_H_
+#ifndef _GLOBAL_H_
+#define _GLOBAL_H_
 
 #include "graphics.h"
 #include "extgraph.h"
@@ -18,55 +18,51 @@
 
 #define MAX_NOTES 10
 #define MAX_TITLE_LENGTH 128
-#define MAX_CATEGORY_LENGTH 128
-
 #define TIMER_ERRORMESSAGE 2
-
-int noteCount = 0;
-int errorCode = 0;
-
-double windowsWidth,windowsHeight;
-double fontHeight;
-
-
-double noteTitleTextboxX;
-double noteTitleTextboxY;
-double noteTitleTextboxW;
-double noteTitleTextboxH;
 
 enum WindowState{
     MAIN_WINDOW,
     NEW_NOTE_WINDOW,
-    EDIT_NOTE_WINDOW,
-    INTRO_WINDOW,
-    FIND_WINDOW
+    EDIT_NOTE_WINDOW
 };
+enum WindowState Window_State;
 
 typedef struct NoteNode NoteNode;
 
 struct NoteNode {
     char title[MAX_TITLE_LENGTH];
-    char category[MAX_CATEGORY_LENGTH];
+    double titleY;
     NoteNode *next;
+    //FILE fp;
 };
+
+//extern struct NoteNode *NoteNode; //struct error
+
+extern double ww,wh;
+extern NoteNode *note_head;
+extern NoteNode *currentNode;
+extern char titleInput[MAX_TITLE_LENGTH];
+extern int noteCount;
+extern int errorCode;
+extern double curyy;
 
 void drawMainWindow();
 void drawNewNoteWindow();
-void drawEditNoteWindow();
+void drawEditNoteWindow(NoteNode *current);
 void drawContent();
 void drawMenu();
 void drawErrorMessage(int errorCode);
 
-void addButtonClick(void);
-void noteTitleClick(int index);
-void createNewNoteWindow(void);
-void saveNewNoteTitle(char *title, char *category);
 void refreshScreen();
 void windowsInit();
 
+void saveNewNoteTitle(char *title);
+void deleteNoteNode(NoteNode *current);
+void noteTitleClick(NoteNode* current);
 
 void KeyboardEventProcess(int key, int event);
 void CharEventProcess(char c);
 void MouseEventProcess(int x, int y, int button, int event);
 void TimerEventProcess(int timerID);
+
 #endif
